chore: fold simple-exchange-frontend into dapp-simple-exchange/ui
