--- packages/dapp-simple-exchange/ui/.env.local
@@ -1,3 +1,0 @@
-REACT_APP_DAPP_API_URL="http://127.0.0.1:8000"
-REACT_APP_WALLET_BRIDGE_URL="http://127.0.0.1:8000"
-REACT_APP_CONTRACT_ID=""
