+++ packages/dapp-simple-exchange/ui/README.md
@@ -1,12 +1,12 @@
-# Autoswap Demo
+# Simple Exchange Demo
 
 ## Running the demo
 
 From the project directory, run:
 
 ```bash
-npm install
-npm start
+yarn
+yarn start
 ```
 
 Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
