+++ packages/dapp-simple-exchange/ui/package.json
@@ -1,19 +1,17 @@
 {
   "name": "@agoric/dapp-simple-exchange-ui",
   "version": "0.0.1",
-  "description": "Frontend UI",
-  "private": true,
+  "description": "Simple Exchange Frontend",
+  "author": "Agoric",
+  "license": "Apache-2.0",
+  "homepage": ".",
   "scripts": {
-    "build": "exit 0",
-    "test": "exit 0",
     "clean": "rimraf ./build",
-    "lint": "eslint './src/**/*.{js,jsx}'",
+    "lint-check": "eslint '**/*.{js,jsx}'",
     "lint-fix": "eslint --fix '**/*.{js,jsx}'",
-    "lint-check": "echo disabling lint in ui until files are lint compliant",
-    "prettier": "prettier './src/**/*.{js,jsx,scss}' --write",
     "start": "react-scripts start",
-    "test-react": "react-scripts test --env=jsdom",
-    "build-react": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "build": "react-scripts build",
     "eject": "react-scripts eject"
   },
   "browserslist": {
@@ -28,28 +26,26 @@
       "last 1 safari version"
     ]
   },
-  "resolutions": {
-    "**/**/serialize-javascript": "^2.1.2"
-  },
   "dependencies": {
-    "@material-ui/core": "^4.2.1",
-    "@material-ui/icons": "^4.5.1",
+    "@material-ui/core": "^4.9.3",
+    "@material-ui/icons": "^4.9.1",
     "clsx": "^1.0.4",
     "http-proxy-middleware": "^1.0.0",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "eslint": "^6.5.1",
+    "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.3.0",
-    "eslint-plugin-import": "^2.18.2",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-eslint-comments": "^3.1.2",
+    "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.15.1",
-    "eslint-plugin-react-hooks": "^2.2.0",
-    "prettier": "^1.18.2",
-    "react-scripts": "^3.2.0",
-    "rimraf": "^3.0.0"
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.18.3",
+    "eslint-plugin-react-hooks": "^2.4.0",
+    "prettier": "^1.19.1",
+    "react-scripts": "^3.4.0",
+    "rimraf": "^3.0.2"
   }
 }
