+++ packages/dapp-simple-exchange/ui/public/index.html
@@ -25,7 +25,7 @@
     -->
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
 
-    <title>AutoSwap Exchange</title>
+    <title>Simple Exchange</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
