+++ packages/dapp-simple-exchange/ui/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "Autoswap",
-  "name": "Autoswap Exchange",
+  "short_name": "Simple Exchange",
+  "name": "Simple Exchange",
   "icons": [
     {
       "src": "favicon.ico",
