+++ packages/dapp-simple-exchange/ui/src/setupProxy.js
@@ -8,7 +8,8 @@ const { createProxyMiddleware } = require('http-proxy-middleware');
 const API_URL = process.env.REACT_APP_DAPP_API_URL;
 const BRIDGE_URL = process.env.REACT_APP_WALLET_BRIDGE_URL;
 
-module.exports = function manualProxy(app) {
+// eslint-disable-next-line func-names
+module.exports = function(app) {
   if (API_URL > '') {
     app.use('/vat', createProxyMiddleware({ target: API_URL }));
   }
