--- packages/dapp-simple-exchange/ui/src/utils/.gitignore
@@ -1,1 +1,0 @@
-contractID.js
