+++ packages/dapp-simple-exchange/ui/src/utils/constants.js
@@ -1,3 +1,3 @@
-export { default as CONTRACT_ID } from './contractID';
+export const CONTRACT_ID = process.env.REACT_APP_CONTRACT_ID;
 export const API_URL = process.env.REACT_APP_DAPP_API_URL;
 export const BRIDGE_URL = process.env.REACT_APP_WALLET_BRIDGE_URL;
