--- packages/simple-exchange-frontend/.env.local.example
@@ -1,3 +1,0 @@
-REACT_APP_DAPP_API_URL=""
-REACT_APP_WALLET_BRIDGE_URL=""
-REACT_APP_CONTRACT_ID=""
