--- packages/simple-exchange-frontend/.eslintrc
@@ -1,35 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "extends": [
-    "airbnb",
-    "plugin:prettier/recommended",
-    "plugin:eslint-comments/recommended"
-  ],
-  "env": {
-    "es6": true
-  },
-  "globals": {
-    "globalThis": false
-  },
-  "rules": {
-    "complexity": ["error", 8],
-    "implicit-arrow-linebreak": "off",
-    "function-paren-newline": "off",
-    "arrow-parens": "off",
-    "strict": "off",
-    "max-depth": ["error", { "max": 4 }],
-    "max-lines": ["error", 300],
-    "max-nested-callbacks": ["error", 4],
-    "max-statements": ["error", 32],
-    "max-statements-per-line": ["error", { "max": 1 }],    "no-console": "off",
-    "no-unused-vars": ["error", { "argsIgnorePattern": "^_" }],
-    "no-return-assign": "off",
-    "no-param-reassign": "off",
-    "no-restricted-syntax": ["off", "ForOfStatement"],
-    "no-unused-expressions": "off",
-    "no-loop-func": "off",
-    "import/extensions": "off",
-    "import/prefer-default-export": "off",
-    "eslint-comments/no-unused-disable": "error"
-  }
-}
