--- packages/simple-exchange-frontend/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "trailingComma": "all",
-  "singleQuote": true
-}
