--- packages/simple-exchange-frontend/README.md
@@ -1,14 +1,0 @@
-# Simple Exchange Demo
-
-## Running the demo
-
-From the project directory, run:
-
-```bash
-yarn
-yarn start
-```
-
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
-
-The page will reload if you make edits.
