--- packages/simple-exchange-frontend/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "Simple Exchange",
-  "name": "Simple Exchange",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "48x48 32x32 16x16",
-      "type": "image/x-icon"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
