--- packages/simple-exchange-frontend/src/index.jsx
@@ -1,13 +1,0 @@
-/* globals document */
-import React from 'react';
-import { render } from 'react-dom';
-
-import ApplicationContextProvider from './contexts/Application';
-import App from './pages/App';
-
-render(
-  <ApplicationContextProvider>
-    <App />
-  </ApplicationContextProvider>,
-  document.getElementById('root'),
-);
