--- packages/simple-exchange-frontend/src/store/actions.js
@@ -1,33 +1,0 @@
-import {
-  ACTIVATE_CONNECTION,
-  DEACTIVATE_CONNECTION,
-  SERVER_CONNECTED,
-  SERVER_DISCONNECTED,
-  UPDATE_PURSES,
-  RESET_STATE,
-} from './types';
-
-export const activateConnection = () => ({
-  type: ACTIVATE_CONNECTION,
-});
-
-export const deactivateConnection = () => ({
-  type: DEACTIVATE_CONNECTION,
-});
-
-export const serverConnected = () => ({
-  type: SERVER_CONNECTED,
-});
-
-export const serverDisconnected = () => ({
-  type: SERVER_DISCONNECTED,
-});
-
-export const updatePurses = purses => ({
-  type: UPDATE_PURSES,
-  payload: purses,
-});
-
-export const resetState = () => ({
-  type: RESET_STATE,
-});
