--- packages/simple-exchange-frontend/src/store/types.js
@@ -1,9 +1,0 @@
-export const ACTIVATE_CONNECTION = 'ACTIVATE_CONNECTION';
-export const DEACTIVATE_CONNECTION = 'DEACTIVATE_CONNECTION';
-
-export const SERVER_CONNECTED = 'SERVER_CONNECTED';
-export const SERVER_DISCONNECTED = 'SERVER_DISCONNECTED';
-
-export const UPDATE_PURSES = 'UPDATE_PURSES';
-
-export const RESET_STATE = 'RESET_STATE';
